requirements.txt
appdirs==1.4.3
APScheduler==3.7.0
asgiref==3.3.1
astroid==2.5.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.8.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
PyJWT==1.7.1
pylint==2.7.4
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
